KPI SBI_KPI_MODEL_INSTANCE : Fix same problems in the update method.
git-svn-id: svn://svn.forge.objectweb.org/svnroot/spagobi@6273 99afaf0d-6903-0410-885a-c66a8bbb5f81

+++ V2.x/trunk/SpagoBIProject/src/it/eng/spagobi/kpi/model/dao/ModelInstanceDAOImpl.java
@@ -261,7 +261,7 @@ public class ModelInstanceDAOImpl extends AbstractHibernateDAO implements
 			}
 			aSession.update(sbiKpiModelInst);
 
-			if (deleteOldHistory) {
+			if (deleteOldHistory && oldSbiKpiInstance != null) {
 				deleteKpiInstance(aSession, oldSbiKpiInstance
 						.getIdKpiInstance());
 			}
